+++ plugins/system/sourcerer/src/Code.php
@@ -1,26 +1,28 @@
 <?php
 /**
  * @package         Sourcerer
- * @version         7.1.4
+ * @version         8.2.0
  * 
  * @author          Peter van Westen <info@regularlabs.com>
  * @link            http://www.regularlabs.com
- * @copyright       Copyright © 2017 Regular Labs All Rights Reserved
+ * @copyright       Copyright © 2019 Regular Labs All Rights Reserved
  * @license         http://www.gnu.org/licenses/gpl-2.0.html GNU/GPL
  */
 
-namespace RegularLabs\Sourcerer;
+namespace RegularLabs\Plugin\System\Sourcerer;
 
 defined('_JEXEC') or die;
 
-use JFactory;
 use JFile;
+use Joomla\CMS\Application\CMSApplication;
+use Joomla\CMS\Factory as JFactory;
+use Joomla\CMS\Version;
 
 class Code
 {
-	public static function run($src_string = '', &$src_variables)
+	public static function run($src_string, &$src_variables, $tmp_path = '')
 	{
-		if (!is_string($src_string) || $src_string == '')
+		if ( ! is_string($src_string) || $src_string == '')
 		{
 			return '';
 		}
@@ -28,11 +30,11 @@ class Code
 		$src_pre_variables = array_keys(get_defined_vars());
 
 		ob_start();
-		$src_post_variables = self::execute($src_string, $src_variables);
+		$src_post_variables = self::execute($src_string, $src_variables, $tmp_path);
 		$src_output         = ob_get_contents();
 		ob_end_clean();
 
-		if (!is_array($src_post_variables))
+		if ( ! is_array($src_post_variables))
 		{
 			return $src_output;
 		}
@@ -54,9 +56,9 @@ class Code
 		return $src_output;
 	}
 
-	private static function execute($string = '', $src_variables)
+	private static function execute($string = '', $src_variables, $tmp_path = '')
 	{
-		$function_name = 'src_' . md5($string);
+		$function_name = 'sourcerer_php_' . md5($string);
 
 		if (function_exists($function_name))
 		{
@@ -65,16 +67,19 @@ class Code
 
 		$contents = self::generateFileContents($function_name, $string);
 
-		$folder    = JFactory::getConfig()->get('tmp_path', JPATH_ROOT . '/tmp');
-		$temp_file = $folder . '/' . $function_name;
+		$tmp_path  = $tmp_path ? $tmp_path : JFactory::getConfig()->get('tmp_path', JPATH_ROOT . '/tmp');
+		$temp_file = $tmp_path . '/' . $function_name;
 
 		JFile::write($temp_file, $contents);
 
 		include_once $temp_file;
 
-		JFile::delete($temp_file);
+		if ( ! defined('JDEBUG') || ! JDEBUG)
+		{
+			JFile::delete($temp_file);
+		}
 
-		if (!function_exists($function_name))
+		if ( ! function_exists($function_name))
 		{
 			// Something went wrong!
 			return [];
@@ -85,9 +90,9 @@ class Code
 
 	private static function generateFileContents($function_name = 'src_function', $string = '')
 	{
-		$init_variables = self::getVarInits();
+		$init = self::getVarInits();
 
-		$init_variables[] =
+		$init[] =
 			'if (is_array($src_variables)) {'
 			. 'foreach ($src_variables as $src_key => $src_value) {'
 			. '${$src_key} = $src_value;'
@@ -98,9 +103,9 @@ class Code
 			'<?php',
 			'defined(\'_JEXEC\') or die;',
 			'function ' . $function_name . '($src_variables){',
-			implode("\n", $init_variables),
-			$string,
-			';return get_defined_vars();',
+			implode("\n", $init),
+			$string . ';',
+			'return get_defined_vars();',
 			';}',
 		];
 
@@ -110,13 +115,43 @@ class Code
 	private static function getVarInits()
 	{
 		return [
-			'$app = $mainframe = JFactory::getApplication();',
-			'$document = $doc = JFactory::getDocument();',
+			'$app = $mainframe = RegularLabs\Plugin\System\Sourcerer\Code::getApplication();',
+			'$document = $doc = RegularLabs\Plugin\System\Sourcerer\Code::getDocument();',
 			'$database = $db = JFactory::getDbo();',
 			'$user = JFactory::getUser();',
 			'$Itemid = $app->input->getInt(\'Itemid\');',
 		];
+	}
 
-		return $init_variables;
+	public static function getApplication()
+	{
+		if (JFactory::getApplication()->input->get('option') != 'com_finder')
+		{
+			return JFactory::getApplication();
+		}
+
+		return CMSApplication::getInstance('site');
+	}
+
+	public static function getDocument()
+	{
+		if (JFactory::getApplication()->input->get('option') != 'com_finder')
+		{
+			return JFactory::getDocument();
+		}
+
+		$lang    = JFactory::getLanguage();
+		$version = new Version;
+
+		$attributes = [
+			'charset'      => 'utf-8',
+			'lineend'      => 'unix',
+			'tab'          => "\t",
+			'language'     => $lang->getTag(),
+			'direction'    => $lang->isRtl() ? 'rtl' : 'ltr',
+			'mediaversion' => $version->getMediaVersion(),
+		];
+
+		return \JDocument::getInstance('html', $attributes);
 	}
 }
